# PERF: Always using panda's hashtable approach, dropping np.in1d