--- a/asv_bench/benchmarks/series_methods.py
+++ b/asv_bench/benchmarks/series_methods.py
@@ -90,10 +90,10 @@ class IsInForObjects:
         self.s_long_floats.isin(self.vals_long_floats)
 
 
-class IsInLongSeries:
+class IsInLongSeriesLookUpDominates:
     params = [
         ["int64", "int32", "float64", "float32", "object"],
-        [1, 2, 5, 10, 50, 100, 1000, 10 ** 5],
+        [1, 2, 5, 10, 16, 50, 100, 1000, 10 ** 5],
         ["random_hits", "random_misses", "monotone_hits", "monotone_misses"],
     ]
     param_names = ["dtype", "MaxNumber", "series_type"]
@@ -114,6 +114,28 @@ class IsInLongSeries:
         self.values = np.arange(MaxNumber).astype(dtype)
 
     def time_isin(self, dtypes, MaxNumber, series_type):
+        self.series.isin(self.values)
+
+
+class IsInLongSeriesValuesDominate:
+    params = [
+        ["int64", "int32", "float64", "float32", "object"],
+        ["random", "monotone"],
+    ]
+    param_names = ["dtype", "series_type"]
+
+    def setup(self, dtype, series_type):
+        N = 10 ** 7
+        if series_type == "random":
+            np.random.seed(42)
+            vals = np.random.randint(0, 10 * N, N)
+        if series_type == "monotone":
+            vals = np.arange(N)
+        self.values = vals.astype(dtype)
+        M = 10 ** 6 + 1
+        self.series = Series(np.arange(M)).astype(dtype)
+
+    def time_isin(self, dtypes, series_type):
         self.series.isin(self.values)