--- a/pandas/core/algorithms.py
+++ b/pandas/core/algorithms.py
@@ -435,11 +435,18 @@ def isin(comps: AnyArrayLike, values: AnyArrayLike) -> np.ndarray:
 
     f = htable.ismember_object
 
-    # an alternative is to use np.in1d if values has a few
-    # elements (about 10) - it is faster than a hash-table
-    # for these cases. However, one must be cautious with
-    # nans (see GH22205)
-    if is_integer_dtype(comps.dtype):
+    # GH16012
+    # Ensure np.in1d doesn't get object types or it *may* throw an exception
+    # Albeit hashmap has O(1) look-up (vs. O(logn) in sorted array),
+    # in1d is faster for small sizes
+    if len(comps) > 1_000_000 and len(values) <= 16 and not is_object_dtype(comps):
+        # If the the values include nan we need to check for nan explicitly
+        # since np.nan it not equal to np.nan
+        if isna(values).any():
+            f = lambda c, v: np.logical_or(np.in1d(c, v), np.isnan(c))
+        else:
+            f = np.in1d
+    elif is_integer_dtype(comps.dtype):
         try:
             values = values.astype("int64", copy=False)
             comps = comps.astype("int64", copy=False)