- [x] tests added / passed
- [x] passes `black pandas`
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry

numpy's `in1d` uses look-up in `O(log n)` in `in1d` compared to panda's `O(1)` so for a large number of unique elements in `values` pandas solution will become better.

Here is comparison in the performance benchmark:

```
       before           after         ratio
     [27aae225]       [796a7dbc]
+     11.4±0.07ms       48.8±0.3ms     4.29  series_methods.IsInLongSeries.time_isin('float64', 1)
+      19.4±0.2ms       59.5±0.6ms     3.06  series_methods.IsInLongSeries.time_isin('float64', 2)
+      13.1±0.1ms       39.2±0.5ms     2.98  series_methods.IsInLongSeries.time_isin('int64', 1)
+      26.9±0.4ms       64.0±0.7ms     2.38  series_methods.IsInLongSeries.time_isin('float32', 1)
+      34.5±0.2ms       74.9±0.9ms     2.17  series_methods.IsInLongSeries.time_isin('float32', 2)
+      23.2±0.1ms       44.5±0.3ms     1.92  series_methods.IsInLongSeries.time_isin('int64', 2)
+      28.1±0.4ms       53.8±0.4ms     1.91  series_methods.IsInLongSeries.time_isin('int32', 1)
+      42.7±0.2ms       70.3±0.3ms     1.64  series_methods.IsInLongSeries.time_isin('float64', 5)
+      38.2±0.3ms       59.1±0.5ms     1.55  series_methods.IsInLongSeries.time_isin('int32', 2)
+      57.4±0.3ms       85.3±0.5ms     1.49  series_methods.IsInLongSeries.time_isin('float32', 5)
-        83.2±1ms       76.6±0.5ms     0.92  series_methods.IsInLongSeries.time_isin('float64', 10)
-         118±6ms       69.8±0.2ms     0.59  series_methods.IsInLongSeries.time_isin('int32', 10)
-         102±1ms       54.9±0.3ms     0.54  series_methods.IsInLongSeries.time_isin('int64', 10)
-         515±3ms          216±1ms     0.42  series_methods.IsInLongSeries.time_isin('int32', 100000)
-         498±3ms          201±1ms     0.40  series_methods.IsInLongSeries.time_isin('int64', 100000)
-         529±2ms          154±7ms     0.29  series_methods.IsInLongSeries.time_isin('float32', 100000)
-         529±3ms        131±0.7ms     0.25  series_methods.IsInLongSeries.time_isin('float64', 100000)
-         519±5ms        107±0.6ms     0.21  series_methods.IsInLongSeries.time_isin('int32', 1000)
-         499±3ms       91.3±0.3ms     0.18  series_methods.IsInLongSeries.time_isin('int64', 1000)
-         534±4ms         87.6±3ms     0.16  series_methods.IsInLongSeries.time_isin('float32', 1000)
-         519±2ms       72.6±0.9ms     0.14  series_methods.IsInLongSeries.time_isin('float64', 1000)
```

It looks as if for more than 10 `values`-elements, panda's approach is faster.

See also this comment (https://github.com/pandas-dev/pandas/issues/22205#issuecomment-410519409).

The question is whether it really makes sense to keep numpy's approach for less than 10 `values`-elements.