@WillAyd 2020-09-24
Nice find. Not sure if we should just lower the limit in the condition to keep performance for a low number of elements to check, but I *think* this is OK