@realead 2020-09-25
@jreback 

I have run all benchmarks which used `isin`, which resulted in

```
asv continuous -f 1.01 upstream/master HEAD -b ^series_methods -b ^categorial
```

The only tests  with  difference were:

```
+     4.44±0.03ms      4.77±0.08ms     1.08  series_methods.NanOps.time_func('var', 1000000, 'int8')
+        961±10μs      1.01±0.02ms     1.05  series_methods.NanOps.time_func('mean', 1000000, 'int8')
-     2.12±0.03ms      1.55±0.08ms     0.73  series_methods.NanOps.time_func('prod', 1000000, 'int8')
-     1.67±0.01ms       1.10±0.1ms     0.66  series_methods.NanOps.time_func('sum', 1000000, 'int8')
```

which I would say is an improvement in overall.

However, I have also changed my introduced benchmarks because they had a bias in the first version: most of the elements from the series were not in the `values`, which has interesting implications:

   * for numpy's `in1d`, which uses binary search as look-up strategy, that means there were almost no cache misses (which more or less is the bottle neck for big tables):  most of the time we search for an element which is larger than the largest element in `values`, thus we always go the same path during the binary search which stays hot in the cache and there are almost no cache misses at all.
  * for khash-table it should make no difference, but it does (and that is a strange thing).

Now the benchmark have another bias: every element from the series is present in `values`, however elements of the series are random thus we can expect a lot of cache misses in numpy's `in1d`, the timing with this "improved" dataset are:

```
       before           after         ratio
     [e4086ec1]       [b93b1e46]
+      18.4±0.2ms        116±0.3ms     6.29  series_methods.IsInLongSeries.time_isin('float64', 2)
+      42.2±0.6ms          195±1ms     4.61  series_methods.IsInLongSeries.time_isin('float64', 5)
+      10.7±0.2ms       46.4±0.1ms     4.33  series_methods.IsInLongSeries.time_isin('float64', 1)
+      33.9±0.3ms        131±0.7ms     3.86  series_methods.IsInLongSeries.time_isin('float32', 2)
+      57.2±0.4ms          210±2ms     3.67  series_methods.IsInLongSeries.time_isin('float32', 5)
+      12.7±0.1ms       42.7±0.4ms     3.37  series_methods.IsInLongSeries.time_isin('int64', 1)
+      80.7±0.3ms        248±0.5ms     3.07  series_methods.IsInLongSeries.time_isin('float64', 10)
+      95.6±0.4ms          263±3ms     2.75  series_methods.IsInLongSeries.time_isin('float32', 10)
+      25.6±0.1ms       62.1±0.7ms     2.42  series_methods.IsInLongSeries.time_isin('float32', 1)
+      28.1±0.1ms         58.3±3ms     2.08  series_methods.IsInLongSeries.time_isin('int32', 1)
+      22.7±0.1ms       42.5±0.2ms     1.88  series_methods.IsInLongSeries.time_isin('int64', 2)
+      37.5±0.5ms       57.1±0.4ms     1.52  series_methods.IsInLongSeries.time_isin('int32', 2)
-        67.1±2ms       57.4±0.5ms     0.86  series_methods.IsInLongSeries.time_isin('int32', 5)
-         1.67±0s       1.26±0.01s     0.76  series_methods.IsInLongSeries.time_isin('float32', 1000)
-      1.66±0.01s          1.25±0s     0.75  series_methods.IsInLongSeries.time_isin('float64', 1000)
-       117±0.7ms       57.4±0.4ms     0.49  series_methods.IsInLongSeries.time_isin('int32', 10)
-       100±0.5ms       42.4±0.2ms     0.42  series_methods.IsInLongSeries.time_isin('int64', 10)
-      2.02±0.02s          286±1ms     0.14  series_methods.IsInLongSeries.time_isin('float32', 100000)
-         2.01±0s        271±0.9ms     0.13  series_methods.IsInLongSeries.time_isin('float64', 100000)
-         1.89±0s       75.5±0.2ms     0.04  series_methods.IsInLongSeries.time_isin('int32', 100000)
-         1.55±0s       57.1±0.3ms     0.04  series_methods.IsInLongSeries.time_isin('int32', 1000)
-         1.87±0s       60.8±0.3ms     0.03  series_methods.IsInLongSeries.time_isin('int64', 100000)
-      1.52±0.01s         42.8±1ms     0.03  series_methods.IsInLongSeries.time_isin('int64', 1000) 
```
Interesting difference to the original benchmark:

 * for  many `values` numpy's `in1d` becomes 3-4 times slower compared to the original benchmark - the cost of more cache misses.
 * while for `int32` and `int64` the runtime of khash-table (panda's approach) is not really dependent on the number of elements in `values` - as expected, this is not the case for `float32` and `float64`.
 * Curiously, `series_methods.IsInLongSeries.time_isin('float64', 100000)`, (i.e. `values` has 10^5 unique elements), is 5 times faster than `series_methods.IsInLongSeries.time_isin('float64', 1000)`, (i.e. `values` has 1000 unique elements): 270ms vs 1250ms.

My assumption is that the hash-functions for float64  (https://github.com/pandas-dev/pandas/blob/2dadd0f48949182dda412eb4032b5234c43c83c1/pandas/_libs/src/klib/khash_python.h#L24) isn't that great and 1.0, 2.0, 3.0 and so on are too often mapped onto the same hash-value. But this needs further inverstigation.