@WillAyd 2020-09-25
> My assumption is that the hash-functions for float64 (

See also #28303