@realead 2020-09-26
Just to illustrate how bad/severe the issue is. Here is the currently used hash-function for `float64` as  a standalone function:

```
%%cython
cdef extern from *:
    """
    #include <string.h>

    inline int64_t f64_to_i64(double val){
          unsigned long long int res; 
          memcpy(&res, &val, sizeof(double)); 
          return res;
    } 


    #define kh_float64_hash_func(key) (unsigned int)((f64_to_i64(key))>>33^(f64_to_i64(key))^(f64_to_i64(key))<<11)
    """
    unsigned int kh_float64_hash_func(double val)
    
def hash_f64(double val):
    return kh_float64_hash_func(val)
```

For 1000 keys, khash set/map will use 2048-buckets, thus we need to see the hashes modulo 2048:

```
n_buckets = 2048
f64hashes=[float64_hash(float(x))%n_buckets for x in range(1000)]
len(set(f64hashes)) # 2
```

gives us only 2 different hashes - 0  and 1024, (for float32 it is even only one (0), but for int32/int64 there are 1000 different!) which mean that the look-up is O(n^2).